folium/requirements.txt
branca==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.0
et-xmlfile==1.1.0
folium==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.5
